Conflicts:
	inc/cache.h
	inc/tracereader.h
	inc/vmem.h
	src/cache.cc
	src/vmem.cc
	test/410-queue-collision.cc
	test/420-issue-prefetch.cc
	test/600-ptw-path.cc

+++ test/modules/prefetcher/address_collector/address_collector.cc
@@ -5,8 +5,8 @@
 
 namespace test
 {
-  std::map<CACHE*, std::vector<uint32_t>> address_operate_collector;
-  std::map<CACHE*, std::vector<uint32_t>> address_fill_collector;
+  std::map<CACHE*, std::vector<uint64_t>> address_operate_collector;
+  std::map<CACHE*, std::vector<uint64_t>> address_fill_collector;
 }
 
 void CACHE::prefetcher_initialize() {}
